+++ lib/browser/rpc-server.ts
@@ -1,29 +1,8 @@
-import { app } from 'electron/main';
-import type { WebContents } from 'electron/main';
 import { clipboard } from 'electron/common';
 import * as fs from 'fs';
 import { ipcMainInternal } from '@electron/internal/browser/ipc-main-internal';
 import * as ipcMainUtils from '@electron/internal/browser/ipc-main-internal-utils';
 import { IPC_MESSAGES } from '@electron/internal/common/ipc-messages';
-
-import type * as desktopCapturerModule from '@electron/internal/browser/desktop-capturer';
-
-const eventBinding = process._linkedBinding('electron_browser_event');
-
-const emitCustomEvent = function (contents: WebContents, eventName: string, ...args: any[]) {
-  const event = eventBinding.createWithSender(contents);
-
-  app.emit(eventName, event, contents, ...args);
-  contents.emit(eventName, event, ...args);
-
-  return event;
-};
-
-const logStack = function (contents: WebContents, code: string, stack: string) {
-  if (stack) {
-    console.warn(`WebContents (${contents.id}): ${code}`, stack);
-  }
-};
 
 // Implements window.close()
 ipcMainInternal.on(IPC_MESSAGES.BROWSER_WINDOW_CLOSE, function (event) {
@@ -57,22 +36,6 @@ ipcMainUtils.handleSync(IPC_MESSAGES.BROWSER_CLIPBOARD_SYNC, function (event, me
 
   return (clipboard as any)[method](...args);
 });
-
-if (BUILDFLAG(ENABLE_DESKTOP_CAPTURER)) {
-  const desktopCapturer = require('@electron/internal/browser/desktop-capturer') as typeof desktopCapturerModule;
-
-  ipcMainInternal.handle(IPC_MESSAGES.DESKTOP_CAPTURER_GET_SOURCES, async function (event, options: Electron.SourcesOptions, stack: string) {
-    logStack(event.sender, 'desktopCapturer.getSources()', stack);
-    const customEvent = emitCustomEvent(event.sender, 'desktop-capturer-get-sources');
-
-    if (customEvent.defaultPrevented) {
-      console.error('Blocked desktopCapturer.getSources()');
-      return [];
-    }
-
-    return await desktopCapturer.getSourcesImpl(event.sender, options);
-  });
-}
 
 const getPreloadScript = async function (preloadPath: string) {
   let preloadSrc = null;
