+++ shell/common/options_switches.cc
@@ -241,8 +241,6 @@ const char kAppUserModelId[] = "app-user-model-id";
 // The application path
 const char kAppPath[] = "app-path";
 
-const char kEnableApiFilteringLogging[] = "enable-api-filtering-logging";
-
 // The command line switch versions of the options.
 const char kScrollBounce[] = "scroll-bounce";
